♻️  refactor(main.py-and-others): moved functions out of main.py and added hydra >>> ⏰ 3h
