--- conf/election.yaml
@@ -1,17 +1,0 @@
-sim:
-  example_name: election
-  gamemasters:
-    online_gamemaster: app_side_only_gamemaster
-    reallife_gamemaster: null
-  load_path: ''
-  model: gpt-4o-mini
-  num_agents: 20
-  num_episodes: 1
-  output_rootname: examples/election/output/N20_T1_Reddit_Big5_independent_v1_news_no_bias_with_images_run1_01_52_15_03_2025/N20_T1_Reddit_Big5_independent_v1_news_no_bias_with_images_run1_01_52_15_03_2025
-  persona_type: Reddit.Big5
-  platform: Mastodon
-  run_name: run1
-  seed: 1
-  sentence_encoder: sentence-transformers/all-mpnet-base-v2
-  use_news_agent: with_images
-  use_server: false
