--- conf/election/probes.yaml
@@ -1,18 +1,0 @@
-probes:
-  queries_data:
-    0:
-      interaction_premise_template:
-        candidate1: Bill Fredrickson
-        candidate2: Bradley Carter
-      query_type: VotePref
-    1:
-      interaction_premise_template:
-        candidate: Bill Fredrickson
-      query_type: Favorability
-    2:
-      interaction_premise_template:
-        candidate: Bradley Carter
-      query_type: Favorability
-    3:
-      query_type: VoteIntent
-  query_lib_module: config_utils.agent_query_lib
